pyproject.toml
[project]
name = "modem-commander"
version = "0.1.0"
description = "Modem configuration automation tool"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.12.13",
    "loguru>=0.7.3",
    "pycryptodome>=3.23.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "telnetlib3>=2.0.4",
]
